+++ built/brave/puppeteer.js
@@ -26,6 +26,9 @@ const disabledBraveFeatures = [
 ];
 const disabledChromeFeatures = [
     'IPH_SidePanelGenericMenuFeature',
+    // Disable because enabling this results in redundant entires in the
+    // MacOS "Local Network" permission table.
+    'MacAppCodeSignClone',
 ];
 const disabledFeatures = disabledBraveFeatures.concat(disabledChromeFeatures);
 const profilePathForArgs = async (args) => {