+++ src/brave/graphml_rewriter.ts
@@ -22,22 +22,6 @@ const attrElmsSelector = 'graphml key'
 const nodeElmsSelector = 'graphml graph node'
 const edgeElmsSelector = 'graphml graph edge'
 
-export const logElm = (elm: Element): void => {
-  let msg = `<${elm.name}`
-  for (const [attrName, attrValue] of Object.entries(elm.attributes)) {
-    msg += ` ${attrName}="${attrValue}"`
-  }
-  msg += '>'
-  if (elm.text) {
-    msg += elm.text
-  }
-  if (elm.children.length > 0) {
-    msg += `\n    - num children: ${elm.children.length}\n`
-  }
-  msg += `</${elm.name}>`
-  console.log(msg)
-}
-
 export class PageGraphXMLRewriter {
   #nodeAttrIds: AttrIdCollection = new Map<AttrName, AttrId>()
   #edgeAttrIds: AttrIdCollection = new Map<AttrName, AttrId>()